Restructure complete
Restructured MQTT demo for the purpose of cert rotation. Reuses the old MQTT demo code to verify the cert is working properly. On startup a good cert is up and working; therefore, on startup everything is the same. The plan is to rotate the cert then reuse this MQTT code with the new cert to verify it's functional.